Moved resize_axis to reader_utils.py
