+++ VideoClassification/reader_utils.py
@@ -2,6 +2,42 @@ import readers
 import tensorflow as tf
 from tensorflow.io import gfile
 
+def resize_axis(tensor, axis, new_size, fill_value=0):
+	"""Truncates or pads a tensor to new_size on on a given axis.
+
+	Truncate or extend tensor such that tensor.shape[axis] == new_size. If the
+	size increases, the padding will be performed at the end, using fill_value.
+
+	Args:
+		tensor: The tensor to be resized.
+		axis: An integer representing the dimension to be sliced.
+		new_size: An integer or 0d tensor representing the new value for
+			tensor.shape[axis].
+		fill_value: Value to use to fill any new entries in the tensor. Will be cast
+			to the type of tensor.
+
+	Returns:
+		The resized tensor.
+	"""
+	tensor = tf.convert_to_tensor(tensor)
+	shape = tf.unstack(tf.shape(tensor))
+
+	pad_shape = shape[:]
+	pad_shape[axis] = tf.maximum(0, new_size - shape[axis])
+
+	shape[axis] = tf.minimum(shape[axis], new_size)
+	shape = tf.stack(shape)
+
+	resized = tf.concat([
+			tf.slice(tensor, tf.zeros_like(shape), shape),
+			tf.fill(tf.stack(pad_shape), tf.cast(fill_value, tensor.dtype))
+	], axis)
+
+	# Update shape.
+	new_shape = tensor.get_shape().as_list()  # A copy is being made.
+	new_shape[axis] = new_size
+	resized.set_shape(new_shape)
+	return resized
 
 def get_reader(feature_names='rgb,audio', feature_sizes='1024,128'):
 	# Convert feature_names and feature_sizes to lists of values.
